requirements.txt
alembic==1.5.8
aniso8601==9.0.1
anyio==2.2.0
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.4
async-generator==1.10
attrs==20.3.0
Babel==2.3.4
backcall==0.2.0
bleach==3.3.0
blinker==1.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
class-registry==2.1.2
Click==7.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.5
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.8
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.2
django-extra-views==0.13.0
django-filter==0.14.0
django-filters-mixin==0.1.4
django-haystack==3.0
django-oscar==3.0.2
django-oscar-api==3.0.0
django-phonenumber-field==3.0.1
django-redis==4.12.1
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
dominate==2.6.0
easy-thumbnails==2.7.1
email-validator==1.1.2
entrypoints==0.3
environ==1.0
factory-boy==2.12.0
Faker==8.1.0
filelock==3.0.12
filters==1.3.2
filters-django==1.0.5
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==2.1.2
flask-ldap3-login==0.9.16
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Principal==0.4.0
flask-restplus==0.9.2
Flask-Security==1.7.5
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.6.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
httpie==2.4.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.18.0
Jinja2==2.11.3
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.13
jupyter-core==4.7.1
jupyter-packaging==0.7.12
jupyter-server==1.5.1
jupyterlab==3.0.12
jupyterlab-pygments==0.1.2
jupyterlab-server==2.4.0
kiwisolver==1.3.1
ldap3==2.9
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.0
mistune==0.8.4
nbclassic==0.2.6
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.20.2
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
passlib==1.7.4
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.10.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
purl==1.5
pyasn1==0.4.8
pycparser==2.20
Pygments==2.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
pysolr==3.9.0
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.0
python-editor==1.0.4
pytube==10.6.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
redis==3.5.3
regex==2021.4.4
requests==2.25.1
requests-toolbelt==0.9.1
Send2Trash==1.5.0
six==1.15.0
sniffio==1.2.0
sorl-thumbnail==12.7.0
sorl-url==0.2.1
sorl-watermark==1.0.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
twilio==6.55.0
urllib3==1.26.4
virtualenv==20.4.3
visitor==0.1.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
Whoosh==2.7.4
WTForms==2.1
